Save converted to separate folder
Initially, the classic nmrPipe strategy of looping through expnos
and locally converting was done. Since this is a serie, now I
save all fids under a folder determining which experiments are selected,
the process all to a separate one.
Headers with the time info are written at the conversion step.
2D slices are concatenated as a pseudo3D time series after processing.