Fix peak picking threshold
Fixed picking thresholds: noise was being calculated but not used
(auto was still input by default because of incorrect flag name)
Changed noise estimation region to 10.5-9.5 ppm / 120 - 110 ppm.
That is OK for IDPs, but the comment is left to use 0% - 100% as
standard. Improvement is not dramatic.
Set threshold to 9*RMSD.
Added a large negative threshold
so negative peaks are not picked